Move coder into its own dir
